backend/ai-services/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
requests==2.32.3
py-algorand-sdk==2.7.0
textblob==0.18.0
deep-translator==1.11.4
python-multipart==0.0.12
python-dotenv==1.0.1
aiofiles==24.1.0
Pillow==10.4.0

